move eval functions to eval_util
